// apps/main/src/models/v2/controlplane/coreModuleVersion.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */
import type { CoreModuleVersionMigrationGeneration } from './coreModuleVersionMigrationGeneration';
import type { ModuleVerificationStatus } from './moduleVerificationStatus';
import type { ModuleVersionSemanticStatus } from './moduleVersionSemanticStatus';

export interface CoreModuleVersion {
  org_id: string;
  module_uuid: string;
  module_slug: string;
  uuid: string;
  semantic_version?: string;
  opaque_version_id: string;
  migration_generation: CoreModuleVersionMigrationGeneration;
  artifact_digest: string;
  verification_status: ModuleVerificationStatus;
  lifecycle_status: ModuleVersionSemanticStatus;
  source_revision: string;
  release_notes?: string;
  /** @minimum 1 */
  resource_version: number;
  published_by?: string;
  created_at: string;
}